requirements.txt
fastmcp
pandas